requirements/common.txt
Django==1.8.18
django_cron==0.5.0
django-appconf==1.0.2
django-countries==5.3.2
django-ipware==2.0.1
django-model-utils==3.1.1
django-pipeline==1.6.14
django-redis==4.8.0
django-softhyphen==1.1.0
django-solo==1.1.3
django-suit==0.2.23
django-suit-ckeditor==0.0.2
beautifulsoup4==4.6.1
html5lib==0.999999999
csscompressor==0.9.4
jsmin==2.2.2
ipython==6.2.1
pillow==4.2.1
psycopg2==2.7.4
pytils==0.3
pytz==2018.3
redis==2.10.6
requests==2.18.4
PyLinkedinAPI==0.1.5
facebook-sdk==2.0.0
python-twitter==3.3
geoip2==2.9.0

# jinja2==2.9.6
# django_jinja==2.3.1
# jinja2-django-tags==0.5

# pysyge==0.2.0
# premailer==3.1.1
# django-mptt==0.8.7
# celery==3.1.23
# django-celery-email==1.1.4
